requirements.txt
unstructured
langchain
openai~=0.27.8
chromadb
tiktoken
python-dotenv~=1.0.0
google-search-results
yfinance~=0.2.28
yahooquery~=2.3.2
matplotlib~=3.7.2
fastapi==0.101.1
uvicorn
streamlit~=1.25.0
pytube~=15.0.0
requests~=2.31.0